Set content length header to 0 on stats post responses.
This is a workaround for what seems like a Varnish bug. Without
the content length header, rails generates an empty response with
transfer-encoding chunked. This is valid but Varnish removes the
transfer encoding header, creating a response without content length
or chunked encoding. This invalid response causes upstream ELBs to
return 502 bad gateway.