Run javascript feature specs only against container
Javascript feature specs pollute the test database because they don't
use transactional fixtures. If FEATURE_SPEC_HOST is set (when running
feature specs against the Docker container) run the Javascript feature
specs with webkit. If it is not set use the rack_test driver for
transactional fixtures and skip the Javascript specs.